// src/interfaces/CsvRowObject.ts
export default interface CsvRowObject {
  KonserttiId: string;
  Aloitusaika: string;
  Esitysjarjestys: string;
  Paivamaara: string;
  Kapellimestari: string;
  Saveltaja: string;
  TeoksenNimi: string;
  TeoksenId: string;
  Sovittaja: string;
  Orkesteri: string;
  Solisti: string;
  KonsertinNimike: string;
  Konserttipaikka: string;
  TietoaKonsertista: string;
  LisatietoaKonsertista: string;
}